+++ credentials/Altium365NexarApi.credentials.ts
@@ -1,34 +1,15 @@
-import type {
-	IAuthenticateGeneric,
-	ICredentialType,
-	INodeProperties,
-} from 'n8n-workflow';
+import type { ICredentialType, INodeProperties } from 'n8n-workflow';
 
 export class Altium365NexarApi implements ICredentialType {
 	name = 'altium365NexarApi';
 
 	displayName = 'Altium 365 Nexar API';
 
-	documentationUrl = 'https://portal.nexar.com/';
+	documentationUrl = 'https://support.nexar.com/support/solutions/articles/101000471994-authorization';
+
+	extends = ['oAuth2Api'];
 
 	properties: INodeProperties[] = [
-		{
-			displayName: 'Client ID',
-			name: 'clientId',
-			type: 'string',
-			default: '',
-			required: true,
-			description: 'The Client ID from your Nexar application',
-		},
-		{
-			displayName: 'Client Secret',
-			name: 'clientSecret',
-			type: 'string',
-			typeOptions: { password: true },
-			default: '',
-			required: true,
-			description: 'The Client Secret from your Nexar application',
-		},
 		{
 			displayName: 'Workspace URL',
 			name: 'workspaceUrl',
@@ -38,12 +19,41 @@ export class Altium365NexarApi implements ICredentialType {
 			required: true,
 			description: 'Your Altium 365 workspace URL',
 		},
-	];
-
-	authenticate: IAuthenticateGeneric = {
-		type: 'generic',
-		properties: {
-			qs: {},
+		{
+			displayName: 'Grant Type',
+			name: 'grantType',
+			type: 'hidden',
+			default: 'authorizationCode',
 		},
-	};
+		{
+			displayName: 'Authorization URL',
+			name: 'authUrl',
+			type: 'hidden',
+			default: 'https://identity.nexar.com/connect/authorize',
+		},
+		{
+			displayName: 'Access Token URL',
+			name: 'accessTokenUrl',
+			type: 'hidden',
+			default: 'https://identity.nexar.com/connect/token',
+		},
+		{
+			displayName: 'Scope',
+			name: 'scope',
+			type: 'hidden',
+			default: 'openid design.domain user.access offline_access',
+		},
+		{
+			displayName: 'Auth URI Query Parameters',
+			name: 'authQueryParameters',
+			type: 'hidden',
+			default: 'access_type=offline',
+		},
+		{
+			displayName: 'Authentication',
+			name: 'authentication',
+			type: 'hidden',
+			default: 'body',
+		},
+	];
 }